UPSTREAM: ath11k: Avoid REO CMD failed prints during firmware recovery
Currently when firmware recovery is in progress, we do not queue REO
commands to the firmware, instead -ESHUTDOWN will be returned to the
caller leading to a failure print on the console. The REO command in
the problem scenario is sent for all tids of a peer in which case we
will have 16 failure prints on the console for a single peer. For an
AP usecase, this count would be even higher in a worst case scenario.
Since these commands are bound to fail during firmware recovery, it
is better to avoid printing these failures and thereby avoid message
flooding on the console.

Tested-on: WCN6750 hw1.0 AHB WLAN.MSL.1.0.1-00887-QCAMSLSWPLZ-1

Fixes: 8ee8d38ca472 ("ath11k: Fix crash during firmware recovery on reo cmd ring access")
Signed-off-by: Manikanta Pubbisetty <quic_mpubbise@quicinc.com>
Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>
Link: https://lore.kernel.org/r/20220602122929.18896-1-quic_mpubbise@quicinc.com
(cherry picked from commit 0ab52b2bd7be8fd49c8ade7703c1faa15359c6c5)

BUG=b:248457431, b:177043467
TEST=emerge-herobrine chromeos-kernel-5_15; deploy and test basic WiFi
     functionalities on herobrine device.

Signed-off-by: Jun Yu <junyuu@chromium.org>
Change-Id: I0dfe692f7d6564f52a86050a0f1f677ca0d7bc59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3943164
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Kevin Lund <kglund@google.com>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>