BACKPORT: FROMGIT: arm64: dts: qcom: sc7180-trogdor: Split out keyboard node and describe detachables
Trogdor devices that have a detachable keyboard still have a
non-detachable keyboard input device present because we include the
cros-ec-keyboard.dtsi snippet in the top-level sc7180-trogdor.dtsi file
that every variant board includes. We do this because the
keyboard-controller node also provides some buttons like the power
button and volume buttons. Unfortunately, this means we register a
keyboard input device that doesn't do anything on boards with a
detachable keyboard.

Change the node's compatible on detachables to the newly introduced
"google,cros-ec-keyb-switches" compatible to indicate that there are
only switches and no keyboard to register. Similarly, move the keyboard
include that defines the keyboard-controller node out of
sc7180-trogdor.dtsi to boards that actually have a keyboard so that the
matrix properties are not defined on boards with the switches
compatible. Future boards can either use the include approach or the
node definition approach to describe a keyboard with possible switches
or just some switches.

Cc: Benson Leung <bleung@chromium.org>
Cc: Guenter Roeck <groeck@chromium.org>
Cc: Douglas Anderson <dianders@chromium.org>
Cc: Hsin-Yi Wang <hsinyi@chromium.org>
Cc: "Joseph S. Barrera III" <joebar@chromium.org>
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Link: https://lore.kernel.org/r/20220627212802.3593012-1-swboyd@chromium.org

(cherry picked from commit a10b760b7402563661fa305882b181a75a1d4894
 git://git.kernel.org/pub/scm/linux/kernel/git/qcom/linux.git for-next)

Conflicts:
   arch/arm64/boot/dts/qcom/sc7180-trogdor-lazor.dtsi
...upstream had the same conflict because of a change that landed in
-fixes vs this in -next. See upstream merge commit 817c2f351974
("Merge tag 'qcom-arm64-fixes-for-5.19' into arm64-for-5.20"). Resolve
like upstream did.

BUG=b:233402918
TEST=Keyboard described properly

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: Iac13381b71446fd689a830f1363d1fe8d58ca040
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3767941
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>