UPSTREAM: media: videodev2.h: add V4L2_CTRL_FLAG_DYNAMIC_ARRAY
Add a new flag that indicates that this control is a dynamically sized
array. Also document this flag.

Currently dynamically sized arrays are limited to one dimensional arrays,
but that might change in the future if there is a need for it.

The initial use-case of dynamic arrays are stateless codecs. A frame
can be divided in many slices, so you want to provide an array containing
slice information for each slice. Typically the number of slices is small,
but the standard allow for hundreds or thousands of slices. Dynamic arrays
are a good solution since sizing the array for the worst case would waste
substantial amounts of memory.

Acked-by: Nicolas Dufresne <nicolas.dufresne@collabora.com>
Tested-by: Benjamin Gaignard <benjamin.gaignard@collabora.com>
Tested-by: Jernej Skrabec <jernej.skrabec@gmail.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>
(cherry picked from commit 64fe675e999c2c7d753ecaaa1349693c59ce6c11)

BUG=b:261593922
TEST=emerge-arm64-generic chromeos-kernel-5_15

Change-Id: I1c68372487c530ac3178deed5380568402477372
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4329169
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Steve Cho <stevecho@chromium.org>