UPSTREAM: drm/encoder: Introduce drmm_encoder_init
The DRM-managed function to register an encoder is
drmm_encoder_alloc() and its variants, which will allocate the underlying
structure and initialisation the encoder.

However, we might want to separate the structure creation and the encoder
initialisation, for example if the structure is shared across multiple DRM
entities, for example an encoder and a connector.

Let's create an helper to only initialise an encoder that would be passed
as an argument.

Acked-by: Thomas Zimmermann <tzimmermann@suse.de>
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Link: https://lore.kernel.org/r/20220711173939.1132294-4-maxime@cerno.tech
(cherry picked from commit f134c9cd9ce829cc66d4d32c57b76de1aab54fe9)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:265055111
TEST=Build test

Backported using forklift.py

Change-Id: I3559dc188b2f3f6569be9cd9d9a24f3063b929f8