FROMGIT: PCI: qcom: Make all optional clocks optional
The kernel is not a devicetree validator and does not need to re-encode
information which is already available in the devicetree.

This is specifically true for the optional PCIe clocks, some of which
are really interconnect clocks.

Treat also the 2.7.0 optional clocks as truly optional instead of
maintaining a list of clocks per compatible (including two compatible
strings for the two identical controllers on sm8450) just to validate
the devicetree.

Link: https://lore.kernel.org/r/20220714071348.6792-7-johan+linaro@kernel.org
Signed-off-by: Johan Hovold <johan+linaro@kernel.org>
Signed-off-by: Lorenzo Pieralisi <lpieralisi@kernel.org>
Reviewed-by: Rob Herring <robh@kernel.org>
Reviewed-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
Reviewed-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>
Reviewed-by: Brian Masney <bmasney@redhat.com>
Acked-by: Stanimir Varbanov <svarbanov@mm-sol.com>
(cherry picked from commit 014aa3518a5826b88a601f5de867551db5c73855
 git://git.kernel.org/pub/scm/linux/kernel/git/helgaas/pci.git next)

BUG=b:244362435
TEST=NVMe works on herobrine

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: I0bb552ebb63a4d9bbc69062f010c89a4788cd068
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3868066
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Douglas Anderson <dianders@chromium.org>