CHROMIUM: iwl7000: mvm: rs-fw: don't crash on missing channel
This is more of a workaround, with MLO we can get here with
the deflink not assigned. It's not critical right now that
we have this right, so WARN_ON_ONCE() and don't crash until
we can fix this area of the code.

Change-Id: I39d23c76eec16ac49f6ae3a6d5f7652041bde855
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
tested: iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: 1d13d975ebe70e6a3f9bbe5149846b800ac9d10e