CHROMIUM: drm/edid: Make 144 Hz not preferred on Sharp LQ140M1JW46
The Sharp LQ140M1JW46 panel is on the Qualcomm sc7280 CRD reference
board. This panel supports 144 Hz and 60 Hz. In the EDID, the 144 Hz
mode is listed first and thus is marked preferred. The EDID decode I
ran says:

  First detailed timing includes the native pixel format and preferred
  refresh rate.

  ...

  Detailed Timing Descriptors:
    DTD 1:  1920x1080  143.981 Hz  16:9   166.587 kHz  346.500 MHz
                 Hfront   48 Hsync  32 Hback  80 Hpol N
                 Vfront    3 Vsync   5 Vback  69 Vpol N
    DTD 2:  1920x1080   59.990 Hz  16:9    69.409 kHz  144.370 MHz
                 Hfront   48 Hsync  32 Hback  80 Hpol N
                 Vfront    3 Vsync   5 Vback  69 Vpol N

I'm proposing here that the above is actually a bug and that the 60 Hz
mode really should be considered preferred by Linux.

The argument here is that this is a laptop panel and on a laptop we
know power will always be a concern. Presumably even if someone using
this panel wanted to use 144 Hz for some use cases they would only do
so dynamically and would still want the default to be 60 Hz.

Let's change the default to 60 Hz using a standard quirk.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Abhinav Kumar <quic_abhinavk@quicinc.com>
(am from https://patchwork.kernel.org/patch/12925827/)
(also found at https://lore.kernel.org/r/20220721152314.RFC.1.Ie333b3e4aff6e4a5b58c4aa805e030e561be8773@changeid)

NOTE: picking this as a CHROMIUM patch since upstream doesn't like
it. See linked bug for details, but so far carrying this as a CHROMIUM
patch seems better than other options. We presumably will want to
carry this forward as long as we think Qualcomm is using sc7280 CRD
boards which is likely for years. If we ever ended up with userspace
able to pick a more reasonable refresh rate, though, we could drop
it. Luckily it's a very small patch.

BUG=b:239951048
TEST=modetest -M msm

Change-Id: Ie333b3e4aff6e4a5b58c4aa805e030e561be8773
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3777393
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Rob Clark <robdclark@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>