UPSTREAM: usb: dwc3: qcom: fix wakeup implementation
It is the Qualcomm glue wakeup interrupts that may be able to wake the
system from suspend and this can now be described in the devicetree.

Move the wakeup-source property handling over from the core driver and
instead propagate the capability setting to the core device during
probe.

This is needed as there is currently no way for the core driver to query
the wakeup setting of the glue device, but it is the core driver that
manages the PHY power state during suspend.

Also don't leave the PHYs enabled when system wakeup has been disabled
through sysfs.

Fixes: 649f5c842ba3 ("usb: dwc3: core: Host wake up support from system suspend")
Signed-off-by: Johan Hovold <johan+linaro@kernel.org>
Link: https://lore.kernel.org/r/20220804151001.23612-9-johan+linaro@kernel.org
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit e3fafbd8e36530eb015a1bb31273bcf7f7035422)

BUG=b:163818407, b:231594967, b:230464230, b:208417985
TEST=USB wakeup works on trogdor and herobrine

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: I2d77394ecae915694cdfe72fbaae70478d6948cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3983211
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>